Fix a dangling pointer access in pj_free
pj_free() will need to look at the ctx of the projection to deallocate
parameter lists. If we free the ctx before, it'll be a dangling pointer
access there and end with a crash. On the other hand, we must free the
context, as pj_free() will not do this for us.